% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileio.r
\name{append_log}
\alias{append_log}
\title{Append To Log}
\usage{
append_log(..., file, sep = "\\n", fill = FALSE, labels = NULL, append = TRUE)
}
\arguments{
\item{...}{Content to append to file.}

\item{file}{A connection, or a character string naming the file to
print to. If "" (the default), cat prints to the standard output
connection, the console unless redirected by sink.}

\item{sep}{A character vector of strings to append after each
element.}

\item{fill}{A logical or (positive) numeric controlling how the
output is broken into successive lines. If FALSE (default), only
newlines created explicitly by "\\n" are printed. Otherwise,
the output is broken into lines with print width equal to the
option width if fill is TRUE, or the value of fill if this is
numeric. Non-positive fill values are ignored, with a warning.}

\item{labels}{Character vector of labels for the lines printed.
Ignored if fill is FALSE.}

\item{append}{Logical. Only used if the argument file is the name
of file (and not a connection or "|cmd") . If TRUE output will
be appended to file; otherwise, it will overwrite the contents
of file.}
}
\description{
Append lines to file. Essentially a wrapper for \code{cat()} with
specific argument values.
}
\keyword{internal}
