% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.r
\name{auto_actions}
\alias{auto_actions}
\title{Automatic Action Consistency Checking}
\usage{
auto_actions(
  workspace = ".",
  actions = c("Removed", "Installed"),
  from,
  to,
  program = "marsh",
  database = "production",
  email.template
)
}
\arguments{
\item{workspace}{The processing workspace. The log file will
be saved here.}

\item{actions}{A vector of action types to check. Default is to
only check field actions, i.e., "Removed" and "Installed".}

\item{from}{The starting date to check actions.}

\item{to}{The ending date to check actions.}

\item{program}{The program name.}

\item{database}{The database name.}

\item{email.template}{The email template used when sending reports
to staff.}
}
\value{
The path to the generated log file.
}
\description{
Automatically check actions for consistency and send report
via email.
}
