% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.r
\name{auto_forms}
\alias{auto_forms}
\title{Automatic Form Processing}
\usage{
auto_forms(
  local.path,
  forms.dir,
  workspace = ".",
  pattern = ".xlsx$",
  limit = 25L,
  program = "marsh",
  database = "production",
  token = NULL,
  email.template
)
}
\arguments{
\item{local.path}{The local path to the SharePoint directory
containing the form directories.}

\item{forms.dir}{SharePoint directory containing forms.}

\item{workspace}{The processing workspace. The log file will
be saved here.}

\item{pattern}{A regex string to filter the forms listing.
Default is to list all modern Excel files, e.g. \code{".xlsx"}.}

\item{limit}{Upper limit to number of forms to process. Limit
applies separately to staged forms and corrected forms.}

\item{program}{The program name.}

\item{database}{The database name.}

\item{token}{The token string. If \code{NULL}, form insert will be
skipped.}

\item{email.template}{The email template used when sending reports
to staff.}
}
\value{
The path to the generated log file.
}
\description{
Automatically process forms, log errors, and send report via email.
}
