% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.r
\name{auto_qaqc}
\alias{auto_qaqc}
\title{Automatic QAQC Processing}
\usage{
auto_qaqc(
  workspace = ".",
  cdec.code,
  analyte.name,
  from,
  to,
  program = "marsh",
  database = "production",
  email.template
)
}
\arguments{
\item{workspace}{The processing workspace. The log file will
be saved here.}

\item{cdec.code}{A vector of station CDEC codes. Regular expressions
Can also be used.}

\item{analyte.name}{A vector of analyte names. Regular expressions
Can also be used.}

\item{from}{The starting date to query QAQC flags.}

\item{to}{The ending date to query QAQC flags.}

\item{program}{The program name.}

\item{database}{The database name.}

\item{email.template}{The email template used when sending reports
to staff.}
}
\value{
The path to the generated log file.
}
\description{
Automatically process QAQC flags and send report via email.
}
