% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.r
\name{bind_results}
\alias{bind_results}
\title{Bind Results}
\usage{
bind_results(
  result.list,
  source.dir,
  archived.dir,
  invalid.dir,
  errored.dir,
  local.path = ""
)
}
\arguments{
\item{result.list}{List of outcome tibbles, i.e. outputs of
\code{process_form()}.}

\item{source.dir}{The common source directory of forms,
typically the "Staged" or "Corrected" directory.}

\item{archived.dir}{The directory containg "archived" forms.}

\item{invalid.dir}{The directory containing "invalid" forms.}

\item{errored.dir}{The directory containing "errored" forms.}

\item{local.path}{The local part of the path to the SharePoint
directory.}
}
\value{
A tibble of formatted results.
}
\description{
Helper function to bind results and reformat file paths.
}
\keyword{internal}
