% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.r
\name{check_actions}
\alias{check_actions}
\title{Check Actions Consistency}
\usage{
check_actions(
  actions = c("Removed", "Installed"),
  from,
  to,
  group.by = c("station", "sonde"),
  program,
  database
)
}
\arguments{
\item{actions}{A vector of action types to check. Default is to
only check field actions, i.e., "Removed" and "Installed".}

\item{from}{The starting date to check actions.}

\item{to}{The ending date to check actions.}

\item{group.by}{The grouping to use. Can be either "station" or
"sonde".}

\item{program}{The program name.}

\item{database}{The database name.}
}
\value{
A table of actions. The field "consistent" lists whether
given action is consistent with the previous action (\code{TRUE})
or not (\code{FALSE}).
}
\description{
Check the history of actions by station or sonde for consistency.
}
