% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.r
\name{check_qaqc}
\alias{check_qaqc}
\title{Check Visual QAQC Progress}
\usage{
check_qaqc(cdec.code, analyte.name, from, to, program, database)
}
\arguments{
\item{cdec.code}{A vector of station CDEC codes. Regular expressions
Can also be used.}

\item{analyte.name}{A vector of analyte names. Regular expressions
Can also be used.}

\item{from}{The starting date to query QAQC flags.}

\item{to}{The ending date to query QAQC flags.}

\item{program}{The program name.}

\item{database}{The database name.}
}
\description{
Count the visual QAQC flags for the specified stations and analytes
over a given period.
}
