% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.r
\name{mail_actions}
\alias{mail_actions}
\title{Email Action Review}
\usage{
mail_actions(
  station.results,
  sonde.results,
  from.date,
  to.date,
  template.path,
  log = ""
)
}
\arguments{
\item{station.results}{An outcome tibble, i.e. output of
\code{check_actions()}.}

\item{template.path}{The YAML template to use.}

\item{log}{File to log results to.}
}
\value{
\code{TRUE} if SMTP server returned successful status code,
\code{FALSE} otherwise.
}
\description{
Send email containing action review report.
}
\keyword{internal}
