% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.r
\name{mail_forms}
\alias{mail_forms}
\title{Email Form Processing Results}
\usage{
mail_forms(results, template.path, log = "")
}
\arguments{
\item{results}{An outcome tibble, i.e. output of \code{process_form()}.}

\item{template.path}{The YAML template to use.}

\item{log}{File to log results to.}
}
\value{
\code{TRUE} if SMTP server returned successful status code,
\code{FALSE} otherwise.
}
\description{
Send email containing form processing report.
}
\keyword{internal}
