% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.r
\name{mail_qaqc}
\alias{mail_qaqc}
\title{Email QAQC Review}
\usage{
mail_qaqc(results, from.date, to.date, template.path, log = "")
}
\arguments{
\item{results}{An outcome tibble, i.e. output of \code{check_qaqc()}.}

\item{template.path}{The YAML template to use.}

\item{log}{File to log results to.}
}
\value{
\code{TRUE} if SMTP server returned successful status code,
\code{FALSE} otherwise.
}
\description{
Send email containing QAQC review report.
}
\keyword{internal}
