% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileio.r
\name{move_file}
\alias{move_file}
\title{Move File}
\usage{
move_file(from, to)
}
\arguments{
\item{from}{Source directory.}

\item{to}{Destination directory.}
}
\description{
Wrapper for \code{file.rename()} that creates destination directory if
it does not exist.
}
\keyword{internal}
