% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.r
\name{outcome_tibble}
\alias{outcome_tibble}
\title{Outcome Tibble}
\usage{
outcome_tibble(outcome, moved = NA, logged = NA)
}
\arguments{
\item{outcome}{Character value specifying outcome.}

\item{moved}{\code{TRUE} if file was moved, \code{FALSE} file move failed,
\code{NA} if move was not attempted.}

\item{logged}{\code{TRUE} if log was created, \code{FALSE} if logging
failed , \code{NA} if logging was not attempted.}
}
\value{
A tibble.
}
\description{
Formatted \code{tibble} of processing results.
}
\keyword{internal}
