% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.r
\name{process_form}
\alias{process_form}
\title{Process Form}
\usage{
process_form(
  form,
  overwrite = FALSE,
  process.log = "",
  staged.dir,
  invalid.dir,
  errored.dir,
  archived.dir,
  program,
  database,
  token
)
}
\arguments{
\item{form}{The form to process.}

\item{overwrite}{If \code{TRUE}, attempt to overwrite existing values
in WQP.}

\item{process.log}{Log file path.}

\item{staged.dir}{The directory containing "staged" forms.}

\item{invalid.dir}{The directory to store forms that could not
be validated.}

\item{errored.dir}{The directory to store forms that encountered
errors during processing.}

\item{archived.dir}{The directory to store forms that were
processed successfully.}

\item{program}{The program name.}

\item{database}{The database name.}

\item{token}{The token string. If \code{NULL}, form insert will be
skipped.}
}
\value{
An outcome tibble, i.e. output of  outcome_tibble()`.
}
\description{
Process a form.
}
