% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileio.r
\name{write_file}
\alias{write_file}
\title{Write File}
\usage{
write_file(to, lines)
}
\arguments{
\item{to}{Destination directory.}

\item{lines}{Vector of lines to write.}
}
\description{
Wrapper for \code{file.create()} and \code{writeLines()} that creates
destination directory if it does not exist.
}
\keyword{internal}
